requirements.txt
django==4.2.4 
Pillow==10.0.1
djangorestframework
